Use SMALLINT if the Spark dataType is ShortType

+++ src/main/scala/io/github/spark_redshift_community/spark/redshift/RedshiftJDBCWrapper.scala
@@ -389,7 +389,7 @@ private[redshift] class JDBCWrapper extends Serializable {
           case LongType => "BIGINT"
           case DoubleType => "DOUBLE PRECISION"
           case FloatType => "REAL"
-          case ShortType => "INTEGER"
+          case ShortType => "SMALLINT"
           case ByteType => "SMALLINT" // Redshift does not support the BYTE type.
           case BooleanType => "BOOLEAN"
           case StringType =>
